+++ lib/pylons_sdk.dart
@@ -1,7 +1,7 @@
 /// Library providing tools for writing client applications that interact with
 /// the Pylons blockchain through an IPC connection with a wallet app in
 /// Flutter.
-library pylons_flutter;
+library pylons_sdk;
 
 export 'src/core/error/exceptions.dart';
 export 'src/features/data/models/profile.dart' show Profile;
@@ -15,4 +15,3 @@ export 'src/generated/pylons/recipe.pb.dart';
 export 'src/generated/pylons/trade.pb.dart';
 export 'src/features/helper/dec_string.dart';
 export 'src/generated/cosmos/base/v1beta1/coin.pb.dart' show Coin;
-