+++ lib/src/pylons_wallet/pylons_wallet_dev.dart
@@ -3,30 +3,28 @@ import 'dart:convert';
 import 'dart:developer';
 
 import 'package:dartz/dartz.dart';
-import 'package:pylons_flutter/src/core/constants/strings.dart';
-import 'package:pylons_flutter/src/features/data/models/transaction.dart';
-import 'package:pylons_flutter/src/features/ipc/completers.dart';
-import 'package:pylons_flutter/src/features/ipc/ipc_constants.dart';
-import 'package:pylons_flutter/src/features/ipc/ipc_handler_factory.dart';
-import 'package:pylons_flutter/src/features/models/sdk_ipc_message.dart';
-import 'package:pylons_flutter/src/generated/pylons/cookbook.pb.dart';
-import 'package:pylons_flutter/src/generated/pylons/item.pb.dart';
-import 'package:pylons_flutter/src/generated/pylons/payment_info.pb.dart';
-import 'package:pylons_flutter/src/generated/pylons/recipe.pb.dart';
-import 'package:pylons_flutter/src/generated/pylons/trade.pb.dart';
-import 'package:pylons_flutter/src/features/models/sdk_ipc_response.dart';
+import 'package:pylons_sdk/src/core/constants/strings.dart';
+import 'package:pylons_sdk/src/features/data/models/transaction.dart';
+import 'package:pylons_sdk/src/features/ipc/completers.dart';
+import 'package:pylons_sdk/src/features/ipc/ipc_constants.dart';
+import 'package:pylons_sdk/src/features/ipc/ipc_handler_factory.dart';
+import 'package:pylons_sdk/src/features/models/sdk_ipc_message.dart';
+import 'package:pylons_sdk/src/generated/pylons/cookbook.pb.dart';
+import 'package:pylons_sdk/src/generated/pylons/item.pb.dart';
+import 'package:pylons_sdk/src/generated/pylons/payment_info.pb.dart';
+import 'package:pylons_sdk/src/generated/pylons/recipe.pb.dart';
+import 'package:pylons_sdk/src/generated/pylons/trade.pb.dart';
+import 'package:pylons_sdk/src/features/models/sdk_ipc_response.dart';
 import 'package:uni_links/uni_links.dart';
 import 'package:url_launcher/url_launcher.dart';
-import '../../pylons_flutter.dart';
+import '../../pylons_sdk.dart';
 import '../pylons_wallet_comm_util.dart';
 
 /// The Pylons class is used for the test net .
 class PylonsWalletDevImpl implements PylonsWallet {
   late StreamSubscription _sub;
 
-
   final String host;
-
 
   PylonsWalletDevImpl(this.host) {
     // // Attach a listener to the stream
@@ -43,7 +41,6 @@ class PylonsWalletDevImpl implements PylonsWallet {
 
     // startIPC();
   }
-
 
   Future<void> handleLink(String link) async {
     log(link, name: '[IPCEngine : handleLink]');
@@ -108,8 +105,6 @@ class PylonsWalletDevImpl implements PylonsWallet {
     throw UnimplementedError();
   }
 
-
-
   @override
   Future<Tuple3<Transaction, Profile, Trade>> txPlaceForSale(
       Item item, int price) {
@@ -118,8 +113,7 @@ class PylonsWalletDevImpl implements PylonsWallet {
   }
 
   @override
-  Future<SDKIPCResponse> txUpdateCookbook(
-      Cookbook cookbook) {
+  Future<SDKIPCResponse> txUpdateCookbook(Cookbook cookbook) {
     // TODO: implement txUpdateCookbook
     throw UnimplementedError();
   }
@@ -127,7 +121,6 @@ class PylonsWalletDevImpl implements PylonsWallet {
   @override
   Future<SDKIPCResponse> txUpdateRecipe(Recipe recipe) async {
     throw UnimplementedError();
-
   }
 
   @override
@@ -148,7 +141,8 @@ class PylonsWalletDevImpl implements PylonsWallet {
       PylonsWalletCommUtil.validateRecipe(recipe);
       var key = Strings.TX_CREATE_RECIPE;
 
-      var sdkIPCMessage = SDKIPCMessage(key, jsonEncode(recipe.toProto3Json()), host);
+      var sdkIPCMessage =
+          SDKIPCMessage(key, jsonEncode(recipe.toProto3Json()), host);
 
       recipeCompleter = Completer();
 
@@ -156,8 +150,6 @@ class PylonsWalletDevImpl implements PylonsWallet {
       return response;
     });
   }
-
-
 
   /// Async: Send the provided message over the IPC channel, then retrieve a
   /// response.
@@ -168,12 +160,12 @@ class PylonsWalletDevImpl implements PylonsWallet {
   ///
   /// The string that is eventually retrieved as a response fits the same
   /// format.
-  Future<SDKIPCResponse> sendMessageNew(SDKIPCMessage sdkipcMessage, Completer<SDKIPCResponse> completer) {
+  Future<SDKIPCResponse> sendMessageNew(
+      SDKIPCMessage sdkipcMessage, Completer<SDKIPCResponse> completer) {
     var encodedMessage = sdkipcMessage.createMessage();
     dispatchUniLink('$BASE_UNI_LINK/$encodedMessage');
     return completer.future;
   }
-
 
   /// This method sends the unilink to the wallet app
   /// [Input] : [unilink] is the unilink with data for the wallet app
@@ -188,13 +180,19 @@ class PylonsWalletDevImpl implements PylonsWallet {
   }
 
   @override
-  Future<SDKIPCResponse> txExecuteRecipe({required String cookbookId, required String recipeName, required List<String> itemIds, required int coinInputIndex, required List<PaymentInfo> paymentInfo}) {
+  Future<SDKIPCResponse> txExecuteRecipe(
+      {required String cookbookId,
+      required String recipeName,
+      required List<String> itemIds,
+      required int coinInputIndex,
+      required List<PaymentInfo> paymentInfo}) {
     // TODO: implement txExecuteRecipe
     throw UnimplementedError();
   }
 
   @override
-  Future<SDKIPCResponse> txEnableRecipe(String cookBookId, String recipeId, String version) {
+  Future<SDKIPCResponse> txEnableRecipe(
+      String cookBookId, String recipeId, String version) {
     // TODO: implement txEnableRecipe
     throw UnimplementedError();
   }