+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+Added support for create, update cookbook.
+Added support for create, update, enable and execute recipe.
+
+## 0.0.2
+Removed unused dependencies.
+Updated description